requirements.txt
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.9.2
PyYAML==6.0.2
APScheduler==3.10.4
playwright==1.48.0
tenacity==9.0.0
python-slugify==8.0.4
feedparser==6.0.11
simhash==2.1.2